+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 
 ### Changes
 
+* [#1248](https://github.com/bbatsov/prelude/issues/1248): Add `-m/--no-move-dotemacs` option to the installer to disable the backing of the user's `.emacs`
 * [#1292](https://github.com/bbatsov/prelude/issues/1292): Add `prelude-python-mode-set-encoding-automatically` defcustom inn `prelude-python.el` module with nil default value.
 * [#1278](https://github.com/bbatsov/prelude/issues/1278): Don't disable `menu-bar-mode` unless `prelude-minimalistic-ui` is enabled.
 * [#1277](https://github.com/bbatsov/prelude/issues/1277): Make it possible to disable the creation of `Super`-based keybindings via `prelude-super-keybindings`.